+++ modules/setup/config/load_strategy.py
@@ -5,20 +5,21 @@ import sys
 
 # Files
 from backtesting.strategy import Strategy
+from modules.setup.config import StrategyDefinition
 
 
-def load_strategy_from_config(config) -> 'Strategy':
+def load_strategy_from_config(strategy_definition: StrategyDefinition) -> Strategy:
     """This function loads the correct strategy, a class that inherits from Strategy and
     made by a user, based on the name and folder specified in the config file"""
 
-    strategies_path = get_full_path_to_strategies_folder(config)
-    sys.path.append(strategies_path) # add to path so that __import__ can find it
+    strategies_path = get_full_path_to_strategies_folder(strategy_definition.strategies_directory)
+    sys.path.append(strategies_path)  # add to path so that __import__ can find it
     for pyfile in os.listdir(strategies_path):
         if pyfile[-3:] != ".py":
             continue
         mod = __import__(pyfile[:-3])
-        if hasattr(mod, config['strategy-name']):
-            CustomStrategy = getattr(mod, config['strategy-name'])
+        if hasattr(mod, strategy_definition.strategy_name):
+            CustomStrategy = getattr(mod, strategy_definition.strategy_name)
             check_if_subclass_of_strategy(CustomStrategy)
 
             try:
@@ -30,15 +31,16 @@ def load_strategy_from_config(config) -> 'Strategy':
             return strategy
 
     # in case it is not found
-    print(f"[ERROR] Could not find strategy '{config['strategy-name']}' in the directory '{strategies_path}'.")
+    print(f"[ERROR] Could not find strategy '{strategy_definition.strategy_name}' in the directory '{strategies_path}'.")
     raise SystemExit
 
 
-def get_full_path_to_strategies_folder(config) -> str:
-    stripped_folder = config['strategies-folder'].strip("./")
+def get_full_path_to_strategies_folder(strategies_folder: str) -> str:
+    stripped_folder = strategies_folder.strip("./")
     strategies_path = os.path.join(utils.get_project_root(), stripped_folder)
     if not os.path.exists(strategies_path):
-        print(f"[ERROR] the strategies folder '{config['strategies-folder']}' (expanded to '{strategies_path}') does not exist.")
+        print(
+            f"[ERROR] the strategies folder '{strategies_folder}' (expanded to '{strategies_path}') does not exist.")
         raise SystemExit
     return strategies_path
