+++ modules/setup/setup.py
@@ -1,7 +1,8 @@
 from pandas import DataFrame
 
 from data.datamodule import DataModule
-from modules.setup.config import ConfigModule, print_pairs
+from modules.algo.algo import AlgoModule
+from modules.setup.config import ConfigModule, print_pairs, load_strategy_from_config
 
 
 class SetupModule(object):
@@ -10,6 +11,8 @@ class SetupModule(object):
         self.config = ConfigModule()
 
     def setup(self) -> dict[str, DataFrame]:
-        config = self.config.get()
-        print_pairs(config)
-        DataModule()
+        config_module = ConfigModule()
+        print_pairs(config_module.raw_config)  # TODO fix mixed level of abstraction
+        historical_data = DataModule(config_module).load_historical_data()
+        load_strategy_from_config(config_module.strategy_definition)
+        AlgoModule(historical_data, config_module)
